@@ -1,24 +1,23 @@
 name: flutter_network_library
 description: A new Flutter package project.
 version: 0.0.3
-author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  http: 0.12.2
+  http: ^0.13.4
   provider:
-  
-  hive: 1.4.4
-  hive_flutter: 0.3.1
-  hive_generator: 0.7.1
 
-  path_provider: 1.6.24
+  hive: ^2.0.5
+  hive_flutter: ^1.1.0
+  hive_generator: ^1.1.2
+
+  path_provider: ^2.0.9
 
 dev_dependencies:
   flutter_test:
