@@ -1,8 +1,8 @@
-import 'package:example/config.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_network_library/data_provider.dart';
+import './config.dart';
 
-void main() async{
+void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   await RESTExecutor.initialize(config, domains);
@@ -15,38 +15,29 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-
-
   @override
   Widget build(BuildContext context) {
     return RESTWidget(
-          executor: RESTExecutor(
-              domain: 'appState',
-              label: 'theme'
+        executor: RESTExecutor(domain: 'appState', label: 'theme'),
+        exact: true,
+        builder: (response) {
+          print("inside listenable");
+          return MaterialApp(
+            title: 'Flutter Demo',
+            theme: ThemeData(
+              brightness: (response.value('dark') ?? true)
+                  ? Brightness.dark
+                  : Brightness.light,
+              primarySwatch: Colors.blue,
+              visualDensity: VisualDensity.adaptivePlatformDensity,
             ),
-            exact: true,
-          builder:(response){
-            print("inside listenable");
-            return MaterialApp(
-        title: 'Flutter Demo',
-        theme: ThemeData(
-          brightness: (response.value('dark')??true)?Brightness.dark:Brightness.light,
-          primarySwatch: Colors.blue,
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-        ),
-        home: MyHomePage(title: 'Network Library Example'),
-      );
-          }
-    );
+            home: MyHomePage(title: 'Network Library Example'),
+          );
+        });
   }
 }
 
-  RESTExecutor _changeTheme = RESTExecutor(
-    domain: 'appState',
-    label: 'theme'
-  );
-
-
+RESTExecutor _changeTheme = RESTExecutor(domain: 'appState', label: 'theme');
 
 class MyHomePage extends StatelessWidget {
   MyHomePage({Key key, this.title}) : super(key: key);
@@ -55,75 +46,65 @@ class MyHomePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-
-      RESTExecutor _getData = RESTExecutor(
-    domain: 'api',
-    label: 'list',
-   
-    retryAfterSeconds: 5,
-    
-  );
+    RESTExecutor _getData = RESTExecutor(
+      domain: 'api',
+      label: 'list',
+      retryAfterSeconds: 5,
+    );
     var data = _getData.watch(context);
 
     print(data.statusCode);
 
     return Scaffold(
       appBar: AppBar(
-
         title: Text(title),
       ),
       body: Center(
         child: Column(
-         
           crossAxisAlignment: CrossAxisAlignment.center,
-          children: <Widget>[ 
+          children: <Widget>[
             Text(
               'For Local State Management',
-              style: TextStyle(
-                fontSize:18,
-                height: 2
-              ),
+              style: TextStyle(fontSize: 18, height: 2),
             ),
-            SizedBox(height: 20,),
+            SizedBox(
+              height: 20,
+            ),
             Text(
-                _changeTheme.response.data.toString(),
-              ),
-            SizedBox(height: 20,),
+              _changeTheme.response.data.toString(),
+            ),
+            SizedBox(
+              height: 20,
+            ),
             RaisedButton(
-              onPressed: (){
-                _changeTheme.execute(
-                  data: {
-                    'dark':!(_changeTheme.response.value('dark')??true)
-                  }
-                );
+              onPressed: () {
+                _changeTheme.execute(data: {
+                  'dark': !(_changeTheme.response.value('dark') ?? true)
+                });
               },
-                          child: Text(
+              child: Text(
                 'Change Theme',
               ),
             ),
             Divider(),
             Text(
               'For Network Request Caching',
-              style: TextStyle(
-                fontSize:18,
-                height: 2
-              ),
+              style: TextStyle(fontSize: 18, height: 2),
             ),
-            SizedBox(height: 20,),
-             Text(
-                              data.fetching?
-                              'Loading...':
-                  data.data.toString(),
-                ),
-             
-            SizedBox(height: 20,),
+            SizedBox(
+              height: 20,
+            ),
+            Text(
+              data.fetching ? 'Loading...' : data.data.toString(),
+            ),
+            SizedBox(
+              height: 20,
+            ),
             RaisedButton(
-              onPressed: (){
-                _getData.execute(
-                  
-                );
+              onPressed: () {
+                _getData.execute();
               },
-                          child: Text(
+              child: Text(
                 'Fetch Data',
               ),
             ),
@@ -134,7 +115,7 @@ class MyHomePage extends StatelessWidget {
           ],
         ),
       ),
-       // This trailing comma makes auto-formatting nicer for build methods.
+      // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
